Move CreateToolNode into it's own src file
Created several PostProcess...() functions for common post-processing.
