+++ src/.srcfiles.yaml
@@ -103,6 +103,7 @@ Files:
     nodes/node_decl.cpp         # Contains the declarations for a node (properties, events, etc.)
     nodes/node_init.cpp         # Initialize NodeCreator class
     nodes/node_prop.cpp         # NodeProperty class
+    nodes/tool_creator.cpp      # Functions for creating new nodes from Ribbon Panel
 
     panels/cstm_propgrid.cpp    # Derived wxPropertyGrid class
     panels/code_display.cpp     # Display code in scintilla control
