+++ src/nodes/node.cpp
@@ -585,137 +585,6 @@ Node* Node::CreateNode(GenName name)
     return cur_selection->CreateChildNode(name);
 }
 
-bool Node::CreateToolNode(GenName name)
-{
-    auto new_node = CreateChildNode(name);
-    if (!new_node)
-        return false;
-
-    auto& frame = wxGetFrame();
-
-    if (name == gen_wxDialog || name == gen_PanelForm || name == gen_wxPanel || name == gen_wxPopupTransientWindow)
-    {
-        auto child_node = new_node->CreateChildNode(gen_VerticalBoxSizer);
-        if (child_node)
-        {
-            if (auto prop = child_node->get_prop_ptr(prop_orientation); prop)
-            {
-                prop->set_value("wxVERTICAL");
-                frame.FirePropChangeEvent(prop);
-            }
-            if (auto prop = child_node->get_prop_ptr(prop_var_name); prop)
-            {
-                child_node->ModifyProperty(prop, "parent_sizer");
-                if (child_node->FixDuplicateName())
-                {
-                    frame.FirePropChangeEvent(prop);
-                }
-            }
-
-            frame.SelectNode(new_node);
-        }
-    }
-    else if (name == gen_wxNotebook || name == gen_wxSimplebook || name == gen_wxChoicebook || name == gen_wxListbook)
-    {
-        new_node = new_node->CreateChildNode(gen_BookPage);
-
-        new_node = new_node->CreateChildNode(gen_VerticalBoxSizer);
-        if (new_node)
-        {
-            if (auto prop = new_node->get_prop_ptr(prop_orientation); prop)
-            {
-                prop->set_value("wxVERTICAL");
-                frame.FirePropChangeEvent(prop);
-            }
-            if (auto prop = new_node->get_prop_ptr(prop_var_name); prop)
-            {
-                new_node->ModifyProperty(prop, "page_sizer");
-                if (new_node->FixDuplicateName())
-                {
-                    frame.FirePropChangeEvent(prop);
-                }
-            }
-        }
-    }
-    else if (name == gen_BookPage)
-    {
-        new_node = new_node->CreateChildNode(gen_VerticalBoxSizer);
-        if (new_node)
-        {
-            if (auto prop = new_node->get_prop_ptr(prop_orientation); prop)
-            {
-                prop->set_value("wxVERTICAL");
-                frame.FirePropChangeEvent(prop);
-            }
-            if (auto prop = new_node->get_prop_ptr(prop_var_name); prop)
-            {
-                new_node->ModifyProperty(prop, "page_sizer");
-                if (new_node->FixDuplicateName())
-                {
-                    frame.FirePropChangeEvent(prop);
-                }
-            }
-        }
-    }
-    else if (name == gen_wxWizard)
-    {
-        new_node = new_node->CreateChildNode(gen_wxWizardPageSimple);
-        new_node->CreateChildNode(gen_VerticalBoxSizer);
-    }
-    else if (name == gen_wxWizardPageSimple)
-    {
-        new_node->CreateChildNode(gen_VerticalBoxSizer);
-    }
-    else if (name == gen_wxMenuBar || name == gen_MenuBar)
-    {
-        auto node_menu = new_node->CreateChildNode(gen_wxMenu);
-        if (node_menu)
-            node_menu->CreateChildNode(gen_wxMenuItem);
-    }
-    else if (name == gen_wxToolBar || name == gen_ToolBar)
-    {
-        new_node->CreateChildNode(gen_tool);
-    }
-    else if (name == gen_wxBoxSizer || name == gen_VerticalBoxSizer || name == gen_wxWrapSizer || name == gen_wxGridSizer ||
-             name == gen_wxFlexGridSizer || name == gen_wxGridBagSizer || name == gen_wxStaticBoxSizer ||
-             name == gen_StaticCheckboxBoxSizer || name == gen_StaticRadioBtnBoxSizer)
-    {
-        auto node = new_node->GetParent();
-        ASSERT(node);
-
-        if (auto prop = node->get_prop_ptr(prop_borders); prop)
-        {
-            if (GetAppOptions().get_SizersAllBorders())
-                prop->set_value("wxALL");
-        }
-
-        if (auto prop = node->get_prop_ptr(prop_flags); prop)
-        {
-            if (GetAppOptions().get_SizersExpand())
-                prop->set_value("wxEXPAND");
-        }
-    }
-    else if (name == gen_wxStdDialogButtonSizer || name == gen_wxStaticLine)
-    {
-        if (auto prop = new_node->get_prop_ptr(prop_flags); prop)
-        {
-            prop->set_value("wxEXPAND");
-            frame.FirePropChangeEvent(prop);
-        }
-    }
-    else if (name == gen_wxContextMenuEvent)
-    {
-        auto event = new_node->GetParent()->GetEvent("wxEVT_CONTEXT_MENU");
-        if (event)
-        {
-            event->set_value(new_node->prop_as_string(prop_handler_name));
-        }
-
-        // Create an initial menu item
-        new_node->CreateChildNode(gen_wxMenuItem);
-    }
-    return true;
-}
 
 void Node::ModifyProperty(PropName name, ttlib::cview value)
 {
